:sparkles: feat: 컴포넌트 분리 및 실링 왁스 기능 추가가
