+++ src/components/letterbox/Envelope/EnvelopeText.tsx
@@ -1,0 +1,20 @@
+import { TextContainer, CombinedText, TitleText, DateText } from "./EnvelopeTextStyles";
+
+interface EnvelopeTextProps {
+  displayText: string;
+  title: string;
+  date: string;
+}
+
+const EnvelopeText = ({ displayText, title, date }: EnvelopeTextProps) => {
+  return (
+    <TextContainer>
+      <CombinedText>
+        <span>{displayText}</span> <TitleText>[ {title} ]</TitleText>
+      </CombinedText>
+      <DateText>{date}</DateText>
+    </TextContainer>
+  );
+};
+
+export default EnvelopeText;
