+++ src/components/letterbox/Envelope/EnvelopeTextStyles.ts
@@ -1,0 +1,26 @@
+import styled from "styled-components";
+
+export const TextContainer = styled.div`
+  text-align: center;
+  margin-top: 0.625rem;
+  line-height: 1.4;
+`;
+
+export const CombinedText = styled.div`
+  font-size: 0.75rem;
+  color: #000;
+  display: flex;
+  justify-content: center;
+  gap: 0.3125rem;
+`;
+
+export const TitleText = styled.span`
+  font-size: 0.75rem;
+  font-weight: bold;
+`;
+
+export const DateText = styled.p`
+  font-size: 0.625rem;
+  color: #000000;
+  margin: 0;
+`;
